+++ gradle.properties
@@ -1,4 +1,4 @@
-version=29.5.0
+version=29.5.1
 group=com.linkedin.pegasus
 org.gradle.configureondemand=true
 org.gradle.parallel=true
