+++ r2-core/src/main/java/com/linkedin/r2/transport/http/client/ratelimiter/RampUpRateLimiterImpl.java
@@ -162,6 +162,12 @@ public class RampUpRateLimiterImpl implements RampUpRateLimiter
   // ############################################## Delegation Section ##############################################
 
   @Override
+  public int getPendingTasksCount()
+  {
+    return _asyncRateLimiter.getPendingTasksCount();
+  }
+
+  @Override
   public Rate getRate()
   {
     return _asyncRateLimiter.getRate();
